Backend/requirements.txt
django==4.2.4
djangorestframework
Pillow
django-rest-authtoken
python-dotenv
djangorestframework-simplejwt
drf-yasg
psycopg2
channels 
channels_redis
daphne
djangochannelsrestframework
django-filter
geopy
aiosmtplib
asyncio
asgiref
celery[redis]
redis==3.5.3
